unify duplicated pytests for taxprofiler and rnafusion
