+++ tests/cli/workflow/taxprofiler/test_cli_taxprofiler_config_case.py
@@ -10,39 +10,3 @@ from click.testing import CliRunner
 from cg.cli.workflow.taxprofiler.base import config_case
 from cg.constants import EXIT_SUCCESS
 from cg.models.cg_config import CGConfig
-
-
-def test_without_options(cli_runner: CliRunner, taxprofiler_context: CGConfig):
-    """Test command without options."""
-
-    # WHEN dry running without anything specified
-    result = cli_runner.invoke(config_case, obj=taxprofiler_context)
-
-    # THEN command should not exit successfully
-    assert result.exit_code != EXIT_SUCCESS
-
-    # THEN command log should inform about missing arguments
-    assert "Missing argument" in result.output
-
-
-def test_with_missing_case(
-    cli_runner: CliRunner,
-    taxprofiler_context: CGConfig,
-    caplog: LogCaptureFixture,
-    case_id_does_not_exist: str,
-):
-    """Test command with invalid case to start with."""
-    caplog.set_level(logging.ERROR)
-    # GIVEN case id not in database
-    assert not taxprofiler_context.status_db.get_case_by_internal_id(
-        internal_id=case_id_does_not_exist
-    )
-
-    # WHEN running
-    result = cli_runner.invoke(config_case, [case_id_does_not_exist], obj=taxprofiler_context)
-
-    # THEN command should NOT successfully call the command it creates
-    assert result.exit_code != EXIT_SUCCESS
-
-    # THEN ERROR log should be printed containing invalid case_id
-    assert "could not be found in StatusDB!" in caplog.text
